Add files via upload
Not fully complete to the standard i would normally like as SSL isnt functional. however, given i was working with limited tools and a 2 day deadline with limited testing. it is functional. chatroom will be updated after dealine as im interested to see how secure i can make this server